move rust mod
